+++ gaios/Assets/TableViewCells/AssetDetailHeaderCell.swift
@@ -1,8 +1,0 @@
-import UIKit
-
-class AssetDetailHeaderCell: UITableViewCell {
-
-    @IBOutlet weak var dismissButton: UIButton!
-    @IBOutlet weak var saveButton: UIButton!
-
-}