+++ gaios/Assets/TableViewCells/AssetDetailTableCell.swift
@@ -1,8 +1,0 @@
-import UIKit
-
-class AssetDetailTableCell: UITableViewCell {
-
-    @IBOutlet weak var titleLabel: UILabel!
-    @IBOutlet weak var detailLabel: UILabel!
-
-}