+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Access app settings screen from hardware wallet connecting screen
 - Converted send transaction button to "slide to send" control
 - Update GDK to version 0.50
+- Refactoring asset detail screen
 
 ### Fixed